+++ scripts/moveTruffleConfig.js
@@ -1,30 +1,6 @@
-const fs = require('fs')
+const copyFile = require('./helpers/utils')
 
 const directory = process.argv.slice(2)[0]
-
-const copyFile = (source, target, cb) => {
-  let cbCalled = false
-
-  let rd = fs.createReadStream(source)
-  rd.on('error', function(err) {
-    done(err)
-  })
-  let wr = fs.createWriteStream(target)
-  wr.on('error', err => {
-    done(err)
-  })
-  wr.on('close', ex => {
-    done()
-  })
-  rd.pipe(wr)
-
-  function done(err) {
-    if (!cbCalled) {
-      cb(err)
-      cbCalled = true
-    }
-  }
-}
 
 const copyStrategy = strategy => {
   try {
